pyproject.toml
[tool.poetry]
name = "mongodb_basic_operations"
version = "0.1.0"
description = ""
authors = ["leblancr <rich.leblanc1@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.7.3"
python-dotenv = "^1.0.1"
pymongoexplain = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
